pyproject.toml
[project]
name = "crm-system"
version = "0.1.0"
authors = [{name = "Luis Abhram Mata", email = "luisabhram@gmail.com"}]
description = "A comprehensive customer relationship management (CRM) system built with Django."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.6.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-allauth>=65.13.1",
    "django-csp>=4.0",
    "django-currentuser>=0.9.0",
    "django-money>=3.5.4",
    "django-phonenumber-field[phonenumbers]>=8.4.0",
    "django-redis>=6.0.0",
    "django-tailwind-cli>=4.4.2",
    "django-unfold>=0.72.0",
    "python-decouple>=3.8",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.0",
    "faker>=38.2.0",
]
prod = [
    "gunicorn>=23.0.0",
    "psycopg[c]>=3.3.1",
]
